pyproject.toml
[project]
name = "commento"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "griotold", email = "gotjd9773@naver.com"},
]
dependencies = ["fastapi>=0.121.1", "uvicorn>=0.38.0", "langchain==0.3.3", "langchain-openai==0.2.2", "openai==1.51.2", "python-dotenv==1.0.1", "httpx==0.27.2", "transformers>=4.57.1", "torch>=2.9.1"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
